rename query of  module name  to jobHistory
