+++ publicService/jobhistory/src/main/java/com/webank/wedatasphere/linkis/jobhistory/transitional/TaskStatus.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2019 WeBank
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.webank.wedatasphere.linkis.jobhistory.transitional;
+
+/**
+ * Created by johnnwang on 2019/6/6.
+ */
+public enum TaskStatus {
+    Inited, WaitForRetry, Scheduled, Running, Succeed, Failed, Cancelled, Timeout
+}
